% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dotmap_data.R
\name{aggregate_dotmap_data}
\alias{aggregate_dotmap_data}
\title{Aggregate dotmap data}
\usage{
aggregate_dotmap_data(dm, s = 4)
}
\arguments{
\item{dm}{dm}

\item{s}{s}
}
\description{
Aggregate dotmap data
}
