% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_lower_zooms.R
\name{aggregate_lower_zooms}
\alias{aggregate_lower_zooms}
\title{Aggregate population totales per pixel for lower zoom levels}
\usage{
aggregate_lower_zooms(dm, i = NULL, j = NULL, logfile = NULL,
  pkg = "pkg")
}
\arguments{
\item{dm}{dotmap_meta object}

\item{i}{tile row. If \code{NULL} (default) all rows are processed}

\item{j}{tile column. If \code{NULL} (default) all columns are processed}
}
\description{
Aggregate population totales per pixel for lower zoom levels
}
