% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_area_maps.R
\name{create_area_maps}
\alias{create_area_maps}
\alias{create_area_maps_sec}
\alias{subtract_area_maps}
\title{Create area maps}
\usage{
create_area_maps(dm, i = NULL, j = NULL, primary = TRUE,
  logfile = NULL)

create_area_maps_sec(dm, i = NULL, j = NULL, primary = FALSE,
  logfile = NULL)

subtract_area_maps(dm, i = NULL, j = NULL, logfile = NULL)
}
\arguments{
\item{dm}{dotmap_meta object}

\item{i}{tile row. If \code{NULL} (default) all rows are processed}

\item{j}{tile column. If \code{NULL} (default) all columns are processed}

\item{primary}{plot primary areas (\code{"area1"}), or secundary (\code{"area2"})?}

\item{logfile}{logfile}
}
\description{
Create maps of the available area at the resolution of specific zoom level, divided in tiles according to the arrangement of a lower zoom level.
}
