% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_functions.R
\name{create_dotmap_website}
\alias{create_dotmap_website}
\title{Create dotmap website}
\usage{
create_dotmap_website(dm, localhost)
}
\arguments{
\item{dm}{dm}

\item{localhost}{localhost}
}
\description{
Create dotmap website
}
