% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_region_per_area_pixel.R
\name{determine_region_per_area_pixel}
\alias{determine_region_per_area_pixel}
\title{Determine region per area pixel}
\usage{
determine_region_per_area_pixel(dm, i = NULL, j = NULL,
  logfile = NULL, pkg = "pkg")
}
\arguments{
\item{dm}{dotmap_meta object}

\item{i}{tile row. If \code{NULL} (default) all rows are processed}

\item{j}{tile column. If \code{NULL} (default) all columns are processed}

\item{logfile}{logfile}

\item{dens_ub}{upperbound of the density per class (e.g. urbanization level)}

\item{dens_lb}{lowerbound of the density per class (e.g. urbanization level)}
}
\description{
The maps created with \code{\link{create_area_maps}} are broken down by the maps created with \code{\link{create_region_maps}}. For each pixel in the first maps, the region id from the second maps are determined.
}
