% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pop_to_pixels.R
\name{dist_pop_to_pixels}
\alias{dist_pop_to_pixels}
\title{Distribute population to pixels at the highest zoom level}
\usage{
dist_pop_to_pixels(dm, i = NULL, j = NULL, logfile = NULL,
  bound = TRUE)
}
\arguments{
\item{dm}{dotmap_info object}

\item{i}{tile row. If \code{NULL} (default) all rows are processed}

\item{j}{tile column. If \code{NULL} (default) all columns are processed}

\item{logfile}{logfile}

\item{bound}{logical value that determines whether all dots should be assigned to pixels, even though there are not enough pixels. \code{TRUE} means that the number of dots that are samples is bounded by the number of pixels.}
}
\description{
Distribute population to pixels at the highest zoom level
}
