% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{dotmap-package}
\alias{dotmap-package}
\alias{dotmap}
\title{Dot Maps}
\description{
Dot Maps are maps where spatial data points are visualized as (colored) dots.
}
\author{
Martijn Tennekes \email{mtennekes@gmail.com}
}
\keyword{GIS,}
\keyword{dot}
\keyword{map}
