% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotmap_data.R
\name{dotmap_project}
\alias{dotmap_project}
\title{Create dotmap project}
\usage{
dotmap_project(dir, area1 = NULL, area2 = NULL, region = NULL, vars,
  var_titles = NULL, var_labels = NULL, pop_totals, pop_tables,
  dens_ub = NULL, dens_lb = NULL, bbx = NA, z, z_arr,
  tile_size = 256, transparent = TRUE, crs, settings)
}
\arguments{
\item{dir}{dir}

\item{area1}{area1}

\item{area2}{area2}

\item{region}{region}

\item{pop_totals}{pop_totals}

\item{pop_tables}{pop_tables}

\item{dens_ub}{dens_ub}

\item{dens_lb}{dens_lb}

\item{bbx}{bbx}

\item{z}{z}

\item{z_arr}{z_arr}

\item{tile_size}{tile_size}

\item{transparent}{transparent}

\item{crs}{crs}

\item{settings}{settings}
}
\description{
Create dotmap project
}
