% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_tile_server.R
\name{make_tile_server}
\alias{make_tile_server}
\title{Make tile server}
\usage{
make_tile_server(dm, logfile = NULL)
}
\arguments{
\item{dm}{dotmap_meta object}

\item{logfile}{logfile}
}
\description{
Make tile server from dotmap tiles that were created with plot_dotmap. The tiles are split to \code{dm$tile_size} by \code{dm$tile_size} bitmap images, and arranged according to the standard convention implemented in the Google Maps API.
}
\seealso{
\url{http://www.maptiler.org/google-maps-coordinates-tile-bounds-projection/}
}
