% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotmap.R
\name{plot_dotmap}
\alias{plot_dotmap}
\title{Plot dotmap}
\usage{
plot_dotmap(dm, i = NULL, j = NULL, logfile = NULL)
}
\arguments{
\item{dm}{dotmap_meta object}

\item{i}{rows. By default \code{NULL} which means all rows are taken}

\item{j}{colums. By default \code{NULL} which means all columns are taken}

\item{logfile}{logfile}
}
\description{
Plot dotmap
}
