% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{NLD_age}
\alias{NLD_age}
\alias{NLD_gender}
\alias{NLD_origin}
\title{Population tables}
\usage{
data(NLD_age)

data(NLD_gender)

data(NLD_origin)
}
\description{
Population tables for age, gender, and origin.
}
