% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pop_to_pixels.R
\name{sample_pop_to_pixels}
\alias{sample_pop_to_pixels}
\title{Sample population to pixels at the highest zoom level}
\usage{
sample_pop_to_pixels(dm, i = NULL, j = NULL, logfile = NULL,
  bound = TRUE, pkg = "pkg")
}
\arguments{
\item{dm}{dotmap_info object}

\item{i}{tile row. If \code{NULL} (default) all rows are processed}

\item{j}{tile column. If \code{NULL} (default) all columns are processed}
}
\description{
Sample population to pixels at the highest zoom level
}
