% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{NLD_area1}
\alias{NLD_area1}
\alias{NLD_area2}
\title{Residential land use areas of the Netherlands.}
\usage{
data(NLD_area1)

data(NLD_area2)
}
\description{
Residential land use areas of the Netherlands in two maps: \code{NLD_area1} are areas where people live, \code{NLD_area2} are areas where people might live, e.g. above shops or in industrial areas. For the dot map, \code{NLD_area1} is used as the primary source. When the population density in those areas reaches a certain threshold, \code{NLD_area2} can be used to place the remaining dots.
}
\details{
\code{NLD_area1} Areas in the Netherlands where people live

\code{NLD_area2} Areas in the Netherlands where people might live, e.g. above shops or in industrial areas.
}
