% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_dotmap.R
\name{show_dotmap}
\alias{show_dotmap}
\title{Show dotmap}
\usage{
show_dotmap(dm, localhost = "http://127.0.0.1", show.region = TRUE)
}
\arguments{
\item{dm}{dm}

\item{localhost}{localhost}

\item{show.region}{show.region}
}
\description{
Show dotmap
}
