% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website_functions.R
\name{write_json_data}
\alias{write_json_data}
\title{Write json data to website}
\usage{
write_json_data(dm, localhost)
}
\arguments{
\item{dm}{dm}

\item{localhost}{localhost}
}
\description{
Write json data to website
}
