% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotmap_data.R
\name{dotmap_project}
\alias{dotmap_project}
\title{Create dotmap project}
\usage{
dotmap_project(dir, area1 = NULL, area2 = NULL, region = NULL, vars,
  var_titles = NULL, var_labels = NULL, pop_totals = NULL,
  pop_tables, dens_ub = NULL, dens_lb = NULL, bbx = NA, z, z_arr,
  tile_size = 256, transparent = TRUE, crs, settings,
  title = "Dotmap", region_title = "Borders",
  labels_title = "Labels", dots_text = "", dotmap_attr = "")
}
\arguments{
\item{dir}{dir}

\item{area1}{area1}

\item{area2}{area2}

\item{region}{region}

\item{vars}{vars}

\item{var_titles}{var_titles}

\item{var_labels}{var_labels}

\item{pop_totals}{pop_totals}

\item{pop_tables}{pop_tables}

\item{dens_ub}{dens_ub}

\item{dens_lb}{dens_lb}

\item{bbx}{bbx}

\item{z}{z}

\item{z_arr}{z_arr}

\item{tile_size}{tile_size}

\item{transparent}{transparent}

\item{crs}{crs}

\item{settings}{settings}

\item{title}{titles}

\item{region_title}{region_title}

\item{labels_title}{labels_title}

\item{dots_text}{dots_text}

\item{dotmap_attr}{dotmap_attr}
}
\description{
Create dotmap project
}
