% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HCL.R
\name{get_HCL_colors}
\alias{get_HCL_colors}
\title{Get HCL colors based on triples.}
\usage{
get_HCL_colors(m, H1 = 0, L.lim = c(80, 20), L.delta = 0.65,
  L.w = 10, zf = 0, C.max = 100, transparent = FALSE,
  output = c("colors", "hcl", "rgb"), H.method = c("cat", "div",
  "seq"), H.div = c(240, 20), L.method = c("v1", "v2"),
  C.method = c("triangle", "entropy"), palette = NA)
}
\arguments{
\item{H1}{hue value of first subpopulation}

\item{L.lim}{limits of the luminance, corresponds to pop.lim}

\item{C.max}{maximum chroma}

\item{transparent}{should the background be transparent? If \code{FALSE}, it will be white.}

\item{output}{a subset of: \code{"colors"}, which outputs the color codes, \code{"hcl"}, which outputs the H, C and L parameters, and \code{"rgb"}, which outputs the red, green and blue values.}

\item{x}{matrix of counts, rows are the cases, columns are the groups. Number of columns should be at most three.}

\item{pop.lim}{limits of the population number per pixel, corresponds to L.lim}
}
\description{
Derive colors from the HCL color space based on triples of values. The H and C are
}
