% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dotmap.R
\name{plot_dotmap}
\alias{plot_dotmap}
\title{Plot dotmap}
\usage{
plot_dotmap(dm, i = NULL, j = NULL, z = NULL, logfile = NULL,
  pkg = "pkg")
}
\arguments{
\item{dm}{dotmap_meta object}

\item{delta}{parameter}

\item{w}{parameter}

\item{...}{arguments passed on to get_HCL_colors}
}
\description{
Plot dotmap
}
